Added a weight slider that's active when the expanded weight options menu isn't. It's a custom class inheriting from TrackBar, but with the rail drawn over and ticks hidden. It has values from 0-100k, which become negative on right-clicking the bar.
The accumulator button in the expanded weight menu is fixed and doesn't rely on a stopwatch. Four button slots remain and will be added later. Might add another 1-2 columns to give the weight display a bit more width for larger numbers. The weight label itself has been removed since it seemed unnecessary. Might append 'KG' as a unit of measurement later.

This will probably be about it for the UI side for a while. Might actually add the Bounce button finally tomorrow, as well as a new idea for splitting the block on considerable impact. Adding an asset for the block would be cool also.